fix(access,webhooks,billing): close the device-trust bypass, the webhook halt, and unblock CI
606 tests green (506 modulith + 100 gateway), 0 failures, 0 skipped.

DEVICE TRUST WAS CROSS-TENANT (V51). setDeviceTrust gated on the orgId in the path but took personId
from the request BODY and never checked the two against each other — ApiPermissionEvaluator only ever
compares #orgId and never sees the person you named. And trust itself was one global boolean on the
device, while require_trusted_device is per-org. V31's header already recorded that second half as a
known modelling bug and deferred it to "its own slice"; this is that slice, and the two halves
together were a bypass needing no guessed identifiers at all: self-signup is open and a new org's
creator is OWNER with every permission, so make your own org, bless your own device with your own
ids, and walk through a different tenant's device policy.

Trust is now a grant BY an organization over a device — user_device_trust (device_id, org_id) — and
absence of a row is absence of trust, so a security default fails in the direction it should. The
membership check closes the other half, through a new shared.security.OrgMembership port rather than
organization's own: organization already depends on access, so the direct edge was a cycle and
ModularityTests caught it. Same inversion PersonLookup/OrgLookup already use.

WEBHOOK DELIVERY COULD HALT PLATFORM-WIDE. The claim applied its subscription-liveness filter in the
OUTER join, after an inner `order by … limit ?` that knew nothing about it — so a paused
subscription's rows filled every claim slot with rows the outer filter then discarded. Reproduced:
39,937 healthy deliveries queued, three consecutive polls claiming ZERO. The stuck rows are never
claimed, so never rescheduled, so never dead-lettered — it does not recover. Predicates now sit
inside, before the LIMIT. The arms are also split, because an OR across the two partial indexes
forced a BitmapOr and sorted the whole backlog to pick 50: 16.7 ms / 3412 kB → 2.7 ms / 29 kB.
Postgres rejects FOR UPDATE anywhere in a UNION, so the locking clause sits on the outer select.

The regression test's point is VOLUME: the existing pause test queues one delivery and can never
reach the LIMIT window, which is exactly why this shipped. Verified it fails against the old query.

CI WAS RED FOR AN ENVIRONMENT REASON. Mockito's inline mock maker self-attaches its ByteBuddy agent —
fine on a laptop, refused in the CI container — surfacing as MockitoInitializationException with no
failing assertion to point at. Both builds now pass the agent at JVM start. Reproduce the restriction
locally with -PblockDynamicAgents; without the fix it reproduces the pipeline's 5 failures exactly.

Also: V50's four proven index corrections (claim indexes that lead with their ORDER BY column, plain
indexes for FK cascade coverage — Postgres cannot use a partial index for referential integrity, 500
person deletes went 6,096 ms → 24 ms — and two audit_log indexes no query reads); geo bbox bounds bind
as BigDecimal so the numeric columns are not cast out of their own index; MCP mutations by a machine
key are attributable instead of anonymous; analytics reports serve a TTL-gated mart instead of moving
300,000 rows per request; usage export batches per subscription.

AGENTS.md §4.6's Flyway advice was wrong and would have reddened the suite: a
`-- flyway:executeInTransaction=false` header comment is inert on flyway-core 12.4.0, which reads
script metadata only from a sidecar .conf. Corrected, with the mechanism that does work.

The reference docs described the pre-refactor world — SRS.md still said the tenant key was the
Keycloak org id "never the local surrogate primary key", exactly backwards. Swept.

+++ src/main/java/ug/co/smsone/geo/internal/GeoSearch.java
@@ -1,5 +1,6 @@
 package ug.co.smsone.geo.internal;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 import ug.co.smsone.shared.geo.GeoStamp;
@@ -19,9 +20,21 @@ interface GeoSearch {
     /**
      * Filters, all optional except {@code orgId}. A bounding box applies only when all four corners are
      * present; a null subject axis is unconstrained.
+     *
+     * <p>The four bounds are {@link BigDecimal} <em>because the column is</em> {@code numeric(9,6)}, and
+     * that is not a style preference — it decides whether the query has an index at all. Declared as
+     * {@code Double} they bind as FLOAT8, so Postgres resolves {@code numeric <= float8} by casting the
+     * COLUMN, and the plan reads {@code (geo_stamp.latitude)::double precision >= …}: an expression over
+     * the column, which {@code geo_stamp_bbox_idx (org_id, latitude, longitude)} cannot answer. Measured
+     * on 150k stamps in one org: seq scan, 56.8&nbsp;ms, 7,379 buffers, two parallel workers conscripted
+     * onto what should be an index probe (the cast also wrecks the estimate, 403 rows guessed as ~4)
+     * versus 0.70&nbsp;ms and 466 buffers on the index once the bounds bind as NUMERIC. Keep them
+     * decimal all the way from the {@code bbox} parameter; if a double ever has to be converted, use
+     * {@link BigDecimal#valueOf(double)} and never {@code new BigDecimal(double)} — the latter turns 1.9
+     * into 1.8999999999999999111…, which silently moves the boundary.
      */
     record Query(UUID orgId, String subjectType, String subjectId,
-            Double minLat, Double minLng, Double maxLat, Double maxLng) {
+            BigDecimal minLat, BigDecimal minLng, BigDecimal maxLat, BigDecimal maxLng) {
 
         boolean hasBoundingBox() {
             return minLat != null && minLng != null && maxLat != null && maxLng != null;
